+++ src/main/java/com/robolucha/game/action/GameAction.java
@@ -7,6 +7,6 @@ import com.robolucha.runner.luchador.LuchadorRunner;
 public interface GameAction {
 
 	public String getName();
-	public void run(LinkedHashMap<Long, LuchadorRunner> runners, LuchadorRunner runner) throws Exception;
+	public void run(LinkedHashMap<Integer, LuchadorRunner> runners, LuchadorRunner runner) throws Exception;
 
 }
